before add button change and delete

+++ src/main/java/com/machete/PP_3_1_Spring_Boot/web/userDao/UserDaoImpl.java
@@ -2,22 +2,20 @@ package com.machete.PP_3_1_Spring_Boot.web.userDao;
 
 import com.machete.PP_3_1_Spring_Boot.web.model.User;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 
 
 import java.util.List;
 
 @Repository
-@Transactional(readOnly = false)
 public class UserDaoImpl implements UserDao {
 
     @PersistenceContext
     private EntityManager entityManager;
 
     @Override
-    @Transactional(readOnly = true)
     public List<User> getAllUsers() {
         return entityManager.createQuery("SELECT u FROM User u", User.class).getResultList();
     }
@@ -25,7 +23,9 @@ public class UserDaoImpl implements UserDao {
     @Override
     public void deleteUser(User user) {
         User us = entityManager.find(User.class, user.getId());
-        entityManager.persist(us);
+        if(us == null) {
+            throw new EntityNotFoundException();
+        }
         entityManager.remove(us);
         entityManager.flush();
     }
@@ -40,7 +40,9 @@ public class UserDaoImpl implements UserDao {
     @Override
     public void changeUser(User user) {
         User us = entityManager.find(User.class, user.getId());
-        entityManager.persist(us);
+        if(us == null) {
+            throw new EntityNotFoundException();
+        }
         us.setId(user.getId());
         us.setName(user.getName());
         us.setLastName(user.getLastName());
